pyproject.toml
[tool.poetry]
name = "react-django"
version = "0.1.0"
description = ""
authors = ["SteShi <ste-shi@mail.ru>"]
readme = "README.md"
packages = [{include = "react_django"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.3"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.13.0"
faker = "^15.3.1"
faker-vehicle = "^0.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
